Rename details and columns
