+++ src/components/items/SimpleDetail.svelte
