+++ src/components/sections/SimpleDetails.svelte
@@ -1,11 +1,11 @@
 <script>
-    import ShortDetail from '../items/ShortDetail.svelte';
+    import SimpleDetail from '../items/SimpleDetail.svelte';
     export let details;
   </script>
   
   <div>
     {#each details as { title, subtitle, link, description }}
-      <ShortDetail {title} {subtitle} {link} {description} />
+      <SimpleDetail {title} {subtitle} {link} {description} />
     {/each}
   </div>
