+++ src/components/structural/Section.svelte
@@ -2,7 +2,7 @@
 	import Contact from '../sections/Contact.svelte';
 	import Skills from '../sections/Skills.svelte';
 	import FullDetails from '../sections/FullDetails.svelte';
-	import ShortDetails from '../sections/ShortDetails.svelte';
+	import SimpleDetails from '../sections/SimpleDetails.svelte';
 	export let label;
 	export let details;
 	export let type;
@@ -11,8 +11,8 @@
 <div class="section">
 	<h2 class="section-label">{label}</h2>
 	<hr class="section-underline" />
-	{#if type === 'short-details'}
-		<ShortDetails {label} {details} />
+	{#if type === 'simple-details'}
+		<SimpleDetails {label} {details} />
 	{:else if type === 'full-details'}
 		<FullDetails {label} {details} />
 	{:else if type === 'skills'}
